+++ NEWS.md
@@ -3,3 +3,4 @@
 * Added correct citation.
 * Fixed: `plot.TrajSpeedIntervals` was not passing additional arguments (`...`) to `plot`.
 * Added: functions `TrajDuration`, `TrajMeanVelocity`, `TrajTranslate`.
+* Added `translateToOrigin` parameter to function `TrajsBuild`.